Refactor to separate UI code from core code
•refactored the project to separate UI code from the core fuctionality code.
•  slight refactor of the UI code
• switched to SFSafariViewController to remove depriciated warning on openURL:
• added support for URL Schemes i.e App-Installer://Your_Download_Link

+++ App Installer/App Installer/InstallButton.m
@@ -20,7 +20,8 @@
 
 -(void)setHighlighted:(BOOL)highlighted {
     [UIView animateWithDuration:0.15 animations:^{
-        self.alpha = highlighted ? 0.8 : 1;
+        // be sure to cast to CGFloat here since the resulting double might not fit into CGFloat
+        self.alpha = (CGFloat) (highlighted ? 0.8 : 1);
         self.transform = highlighted ? CGAffineTransformScale(CGAffineTransformIdentity, 0.95, 0.95) : CGAffineTransformIdentity;
     }];
 }
@@ -33,7 +34,8 @@
 -(void)setEnabled:(BOOL)enabled {
     [super setEnabled:enabled];
     [UIView animateWithDuration:0.15 animations:^{
-        self.alpha = enabled ? 1 : 0.5;
+        // be sure to cast to CGFloat here since the resulting double might not fit into CGFloat
+        self.alpha = (CGFloat) (enabled ? 1 : 0.5);
     }];
 }
